+++ tsk/vs/tsk_vs_i.h
@@ -32,7 +32,7 @@ extern TSK_VS_INFO *tsk_vs_gpt_open(TSK_IMG_INFO *, TSK_DADDR_T);
 
 extern uint8_t tsk_vs_part_unused(TSK_VS_INFO *);
 extern TSK_VS_PART_INFO *tsk_vs_part_add(TSK_VS_INFO *, TSK_DADDR_T,
-    TSK_DADDR_T, TSK_VS_PART_FLAG_ENUM, char *, int8_t, int8_t);
+    TSK_DADDR_T, TSK_VS_PART_FLAG_ENUM, char *, int8_t, int16_t);
 extern void tsk_vs_part_free(TSK_VS_INFO *);
 
 // Endian macros - actual functions in misc/
